requirements.txt
mcp[cli]
caldav
python-dotenv
icalendar
python-dateutil
